fix: sql and table in dataset, reorganize files
