+++ src/commands/bim-to-sml/bim-converter/connection-converter.ts
@@ -69,7 +69,6 @@ export class ConnectionConverter {
         );
       }
     });
-    // }
   }
 
   listUsedConnections(result: SmlConverterResult): Set<string> {
